+++ pybillomat/reminders.py
@@ -7,7 +7,6 @@ Mahnungen
 - Deutsche API-Beschreibung: http://www.billomat.com/de/api/mahnungen
 """
 
-import datetime
 import xml.etree.ElementTree as ET
 from bunch import Bunch
 from http import Url
@@ -16,11 +15,11 @@ from _items_base import Item, ItemsIterator
 
 
 def _reminder_xml(
-    invoice_id = None,
-    contact_id = None,
+    invoice_id = None,  # integer
+    contact_id = None,  # integer
     address = None,
-    date = None,
-    due_date = None,
+    date = None,  # date
+    due_date = None,  # date
     subject = None,
     label = None,
     intro = None,
@@ -33,17 +32,17 @@ def _reminder_xml(
     integer_field_names = [
         "invoice_id",
         "contact_id",
-        "",
     ]
-    float_fieldnames = [
-        "",
-        "",
-        "",
+    date_fieldnames = [
+        "date",
+        "due_date",
     ]
     string_fieldnames = [
-        "",
-        "",
-        "",
+        "address",
+        "subject",
+        "label",
+        "intro",
+        "note",
     ]
 
     reminder_tag = ET.Element("reminder")
@@ -56,12 +55,12 @@ def _reminder_xml(
             new_tag.text = unicode(int(value))
             reminder_tag.append(new_tag)
 
-    # Float Fields
-    for field_name in float_fieldnames:
+    # Date fields
+    for field_name in date_fieldnames:
         value = locals()[field_name]
         if value is not None:
             new_tag = ET.Element(field_name)
-            new_tag.text = unicode(float(value))
+            new_tag.text = value.isoformat()
             reminder_tag.append(new_tag)
 
     # String Fields
@@ -76,7 +75,6 @@ def _reminder_xml(
 
     # Finished
     return xml
-
 
 
 class Reminder(Item):
@@ -98,22 +96,20 @@ class Reminder(Item):
 
         self.id = id  # integer
         self.created = None  # datetime
-        self.reminder_number = None
-        self.number = None  # integer
-        self.number_pre = None
-        self.title = None
-        self.description = None
-        self.sales_price = None  # float
-        self.sales_price2 = None  # float
-        self.sales_price3 = None  # float
-        self.sales_price4 = None  # float
-        self.sales_price5 = None  # float
-        self.currency_code = None
-        self.unit_id = None  # integer
-        self.tax_id = None  # integer
-        self.purchase_price = None  # float
-        self.purchase_price_net_gross = None
-        self.supplier_id = None  # integer
+        self.status = None
+        self.invoice_id = None  # integer
+        self.contact_id = None  # integer
+        self.reminder_text_id = None  # integer
+        self.reminder_level = None  # integer
+        self.reminder_level_name = None
+        self.date = None  # date
+        self.label = None
+        self.subject = None
+        self.intro = None
+        self.note = None
+        self.due_date = None  # date
+        self.total_gross = None  # float
+        self.is_old = None
 
         if reminder_etree is not None:
             self.load_from_etree(reminder_etree)
@@ -125,70 +121,48 @@ class Reminder(Item):
     def create(
         cls,
         conn,
-        number_pre = None,
-        number = None,
-        number_length = None,
-        title = None,
-        description = None,
-        sales_price = None,
-        sales_price2 = None,
-        sales_price3 = None,
-        sales_price4 = None,
-        sales_price5 = None,
-        currency_code = None,
-        unit_id = None,
-        tax_id = None,
-        purchase_price = None,
-        purchase_price_net_gross = None,
-        supplier_id = None
+        invoice_id = None,
+        contact_id = None,
+        address = None,
+        date = None,
+        due_date = None,
+        subject = None,
+        label = None,
+        intro = None,
+        note = None
     ):
         """
         Creates a reminder
 
         :param conn: Connection-Object
 
-        :param number_pre: Prefix
-        :param number: Sequential number
-        :param number_length: inimum length of the customer number
-            (to be filled with leading zeros)
-        :param title: Title
-        :param description: Description
-        :param sales_price: Price
-        :param sales_price2: Price for clients which are members of pricegroup 2.
-            The normal price is used if no price is defined.
-        :param sales_price3: Price for clients which are members of pricegroup 3.
-            The normal price is used if no price is defined.
-        :param sales_price4: Price for clients which are members of pricegroup 4.
-            The normal price is used if no price is defined.
-        :param sales_price5: Price for clients which are members of pricegroup 5.
-            The normal price is used if no price is defined.
-        :param currency_code: Currency
-        :param unit_id: ID of the chosen unit
-        :param tax_id: ID of the chosen tax rate
-        :param purchase_price: Purchase price
-        :param purchase_price_net_gross: Price basis of purchase price
-            (gross or net prices). Possible values: "NET", "GROSS"
-        :param supplier_id: ID of the chosen supplier
+        :param invoice_id: ID of the overdue invoice
+        :param contact_id: ID of the contact
+        :param address: the address;
+            Default: Address from invoice
+        :param date: Date of the reminder; Default: today
+        :param due_date: Due date of the reminder;
+            Default: date + due days taken from the settings
+        :param subject: Subject;
+            Default: Subject of the next dunning level (if available)
+        :param label: Label text to describe the project
+        :param intro: Introductory text;
+            Default: Introductory text of the next dunning level (if available)
+        :param note: Explanatory notes;
+            Default: Explanatory notes of the next dunning level (if available)
         """
 
         # XML
         xml = _reminder_xml(
-            number_pre = number_pre,
-            number = number,
-            number_length = number_length,
-            title = title,
-            description = description,
-            sales_price = sales_price,
-            sales_price2 = sales_price2,
-            sales_price3 = sales_price3,
-            sales_price4 = sales_price4,
-            sales_price5 = sales_price5,
-            currency_code = currency_code,
-            unit_id = unit_id,
-            tax_id = tax_id,
-            purchase_price = purchase_price,
-            purchase_price_net_gross = purchase_price_net_gross,
-            supplier_id = supplier_id
+            invoice_id = invoice_id,
+            contact_id = contact_id,
+            address = address,
+            date = date,
+            due_date = due_date,
+            subject = subject,
+            label = label,
+            intro = intro,
+            note = note
         )
 
         # Send POST-request
@@ -208,50 +182,35 @@ class Reminder(Item):
     def edit(
         self,
         id = None,
-        number_pre = None,
-        number = None,
-        number_length = None,
-        title = None,
-        description = None,
-        sales_price = None,
-        sales_price2 = None,
-        sales_price3 = None,
-        sales_price4 = None,
-        sales_price5 = None,
-        currency_code = None,
-        unit_id = None,
-        tax_id = None,
-        purchase_price = None,
-        purchase_price_net_gross = None,
-        supplier_id = None
+        invoice_id = None,
+        contact_id = None,
+        address = None,
+        date = None,
+        due_date = None,
+        subject = None,
+        label = None,
+        intro = None,
+        note = None
     ):
         """
         Edit a reminder
 
         :param id: ID of the reminder
 
-        :param number_pre: Prefix
-        :param number: Sequential number
-        :param number_length: inimum length of the customer number
-            (to be filled with leading zeros)
-        :param title: Title
-        :param description: Description
-        :param sales_price: Price
-        :param sales_price2: Price for clients which are members of pricegroup 2.
-            The normal price is used if no price is defined.
-        :param sales_price3: Price for clients which are members of pricegroup 3.
-            The normal price is used if no price is defined.
-        :param sales_price4: Price for clients which are members of pricegroup 4.
-            The normal price is used if no price is defined.
-        :param sales_price5: Price for clients which are members of pricegroup 5.
-            The normal price is used if no price is defined.
-        :param currency_code: Currency
-        :param unit_id: ID of the chosen unit
-        :param tax_id: ID of the chosen tax rate
-        :param purchase_price: Purchase price
-        :param purchase_price_net_gross: Price basis of purchase price
-            (gross or net prices). Possible values: "NET", "GROSS"
-        :param supplier_id: ID of the chosen supplier
+        :param invoice_id: ID of the overdue invoice
+        :param contact_id: ID of the contact
+        :param address: the address;
+            Default: Address from invoice
+        :param date: Date of the reminder; Default: today
+        :param due_date: Due date of the reminder;
+            Default: date + due days taken from the settings
+        :param subject: Subject;
+            Default: Subject of the next dunning level (if available)
+        :param label: Label text to describe the project
+        :param intro: Introductory text;
+            Default: Introductory text of the next dunning level (if available)
+        :param note: Explanatory notes;
+            Default: Explanatory notes of the next dunning level (if available)
         """
 
         # Parameters
@@ -262,22 +221,15 @@ class Reminder(Item):
 
         # XML
         xml = _reminder_xml(
-            number_pre = number_pre,
-            number = number,
-            number_length = number_length,
-            title = title,
-            description = description,
-            sales_price = sales_price,
-            sales_price2 = sales_price2,
-            sales_price3 = sales_price3,
-            sales_price4 = sales_price4,
-            sales_price5 = sales_price5,
-            currency_code = currency_code,
-            unit_id = unit_id,
-            tax_id = tax_id,
-            purchase_price = purchase_price,
-            purchase_price_net_gross = purchase_price_net_gross,
-            supplier_id = supplier_id
+            invoice_id = invoice_id,
+            contact_id = contact_id,
+            address = address,
+            date = date,
+            due_date = due_date,
+            subject = subject,
+            label = label,
+            intro = intro,
+            note = note
         )
 
         # Path
@@ -289,15 +241,144 @@ class Reminder(Item):
             raise errors.BillomatError(unicode(response.data, encoding = "utf-8"))
 
 
-    # def get_tags(self):
-    #     """
-    #     Gibt eine Liste mit Schlagworten der Mahnung zurück
-    #     """
-    #
-    #     # Parameters
-    #     if not self.id:
-    #         raise errors.NoIdError()
-    #     ...
+    def complete(self, template_id = None):
+        """
+        Closes a reminder in the draft status (DRAFT).
+        Here, the status of open (OPEN) or overdue (Overdue) is set and
+        a PDF is generated and stored in the file system.
+        The optional paramter template_id determines which template is used
+        to create a pdf. If this parameter is not specified, the default
+        template set is used.
+        """
+
+        # Path
+        path = "/api/reminders/{id}/complete".format(id = self.id)
+
+        # XML
+        complete_tag = ET.Element("complete")
+        if template_id:
+            template_id_tag = ET.Element("template_id")
+            template_id_tag.text = str(template_id)
+            complete_tag.append(template_id_tag)
+        xml = ET.tostring(complete_tag)
+
+        # Send PUT-request
+        response = self.conn.put(path = path, body = xml)
+
+        if response.status != 200:
+            # Parse response
+            error_text_list = []
+            for error in ET.fromstring(response.data):
+                error_text_list.append(error.text)
+
+            # Raise Error
+            raise errors.BillomatError("\n".join(error_text_list))
+
+
+    def send(
+        self,
+        from_address = None,
+        to_address = None,
+        cc_address = None,
+        bcc_address = None,
+        subject = None,
+        body = None,
+        filename = None,
+        # attachments = None
+    ):
+        """
+        Sends the reminder per e-mail to the customer
+
+        :param from_address: (originally: from) Sender
+        :param to_address: (originally: recepients)
+        :param cc_address: (originally: recepients)
+        :param bcc_address: (originally: recepients)
+        :param subject: Subject of the e-mail (may include placeholders)
+        :param body: Text of the e-mail (may include placeholders)
+        :param filename: Name of the PDF file (without .pdf)
+        # :param attachments: List with Dictionaries::
+        #
+        #     [
+        #         {
+        #             "filename": "<Filename>",
+        #             "mimetype": "<MimeType>",
+        #             "base64file": "<Base64EncodedFile>"
+        #         },
+        #         ...
+        #     ]
+        """
+
+        # Path
+        path = "{base_path}/{id}/email".format(
+            base_path = self.base_path,
+            id = self.id
+        )
+
+        # XML
+        email_tag = ET.Element("email")
+
+        # From
+        if from_address:
+            from_tag = ET.Element("from")
+            from_tag.text = from_address
+            email_tag.append(from_tag)
+
+        # Recipients
+        if to_address or cc_address or bcc_address:
+            recipients_tag = ET.Element("recipients")
+            email_tag.append(recipients_tag)
+
+            # To
+            if to_address:
+                to_tag = ET.Element("to")
+                to_tag.text = to_address
+                recipients_tag.append(to_tag)
+
+            # Cc
+            if cc_address:
+                cc_tag = ET.Element("cc")
+                cc_tag.text = cc_address
+                recipients_tag.append(cc_tag)
+
+            # Bcc
+            if bcc_address:
+                bcc_tag = ET.Element("bcc")
+                bcc_tag.text = bcc_address
+                recipients_tag.append(bcc_tag)
+
+        # Subject
+        if subject:
+            subject_tag = ET.Element("subject")
+            subject_tag.text = subject
+            email_tag.append(subject_tag)
+
+        # Body
+        if body:
+            body_tag = ET.Element("body")
+            body_tag.text = body
+            email_tag.append(body_tag)
+
+        # Filename
+        if filename:
+            filename_tag = ET.Element("filename")
+            filename_tag.text = filename
+            filename_tag.append(filename_tag)
+
+        # ToDo: Attachments
+
+        xml = ET.tostring(email_tag)
+
+        # Send POST-request
+        response = self.conn.post(path = path, body = xml)
+
+        if response.status != 200:
+            # Parse response
+            error_text_list = []
+            for error in ET.fromstring(response.data):
+                error_text_list.append(error.text)
+
+            # Raise Error
+            raise errors.BillomatError("\n".join(error_text_list))
 
 
 class Reminders(list):
@@ -321,13 +402,18 @@ class Reminders(list):
     def search(
         self,
         # Search parameters
-        article_number = None,
-        title = None,
-        description = None,
-        currency_code = None,
-        unit_id = None,
+        client_id = None,
+        contact_id = None,
+        invoice_number = None,
+        status = None,
+        from_date = None,
+        to_date = None,
+        subject = None,
+        label = None,
+        intro = None,
+        note = None,
         tags = None,
-        supplier_id = None,
+        article_id = None,
 
         order_by = None,
         fetch_all = False,
@@ -341,13 +427,20 @@ class Reminders(list):
 
         If no search criteria given --> all reminders will returned (REALLY ALL!).
 
-        :param article_number: Reminder number
-        :param title: Title
-        :param description: Description
-        :param currency_code: ISO code of the currency
-        :param unit_id: ID of the chosen unit
+        :param client_id: ID of the client
+        :param contact_id: ID of the contact
+        :param invoice_number: Number of the related invoice
+        :param status: Status (DRAFT, OPEN, PAID, OVERDUE, CANCELED).
+            More than one statuses could be given as a comma separated list.
+            Theses statuses will be logically OR-connected.
+        :param from_date: Only show invoices since this date
+        :param to_date: Only show invoices up to this date
+        :param subject: Free text search in subject
+        :param label: Free text search in label text
+        :param intro: Free text search in introductory text
+        :param note: Free text search in explanatory notes
         :param tags: Comma seperated list of tags
-        :param supplier_id: ID of the chosen supplier
+        :param article_id: ID of an article
 
         :param order_by: Sortings consist of the name of the field and
             sort order: ASC for ascending resp. DESC for descending order.
@@ -356,19 +449,24 @@ class Reminders(list):
             comma.
 
         :param allow_empty_filter: If `True`, every filter-parameter may be empty.
-            So, all reminders will returned. !!! EVERY INVOICE !!!
+            So, all reminders will returned. !!! EVERY !!!
         """
         
         # Check empty filter
         if not allow_empty_filter:
             if not any([
-                article_number,
-                title,
-                description,
-                currency_code,
-                unit_id,
+                client_id,
+                contact_id,
+                invoice_number,
+                status,
+                from_date,
+                to_date,
+                subject,
+                label,
+                intro,
+                note,
                 tags,
-                supplier_id,
+                article_id,
             ]):
                 raise errors.EmptyFilterError()
 
@@ -389,22 +487,30 @@ class Reminders(list):
             url.query["order_by"] = order_by
 
         # Search parameters
-        if article_number:
-            url.query["article_number"] = article_number
-        if title:
-            url.query["title"] = title
-        if article_number:
-            url.query["article_number"] = article_number
-        if description:
-            url.query["description"] = description
-        if currency_code:
-            url.query["currency_code"] = currency_code
-        if unit_id:
-            url.query["unit_id"] = unit_id
+        if client_id:
+            url.query["client_id"] = client_id
+        if contact_id:
+            url.query["contact_id"] = contact_id
+        if invoice_number:
+            url.query["invoice_number"] = invoice_number
+        if status:
+            url.query["status"] = status
+        if from_date:
+            url.query["from_date"] = from_date.isoformat()
+        if to_date:
+            url.query["to_date"] = to_date.isoformat()
+        if subject:
+            url.query["subject"] = subject
+        if label:
+            url.query["label"] = label
+        if intro:
+            url.query["intro"] = intro
+        if note:
+            url.query["note"] = note
         if tags:
             url.query["tags"] = tags
-        if supplier_id:
-            url.query["supplier_id"] = supplier_id
+        if article_id:
+            url.query["article_id"] = article_id
 
         # Fetch data
         response = self.conn.get(path = str(url))
@@ -428,13 +534,18 @@ class Reminders(list):
         if fetch_all and self.total > (self.page * self.per_page):
             self.search(
                 # Search parameters
-                article_number = article_number,
-                title = title,
-                description = description,
-                currency_code = currency_code,
-                unit_id = unit_id,
+                client_id = client_id,
+                contact_id = contact_id,
+                invoice_number = invoice_number,
+                status = status,
+                from_date = from_date,
+                to_date = to_date,
+                subject = subject,
+                label = label,
+                intro = intro,
+                note = note,
                 tags = tags,
-                supplier_id = supplier_id,
+                article_id = article_id,
 
                 order_by = order_by,
                 fetch_all = fetch_all,
@@ -459,26 +570,37 @@ class RemindersIterator(ItemsIterator):
         self.items = Reminders(self.conn)
         self.per_page = per_page
         self.search_params = Bunch(
-            article_number = None,
-            title = None,
-            description = None,
-            currency_code = None,
-            unit_id = None,
+            client_id = None,
+            contact_id = None,
+            invoice_number = None,
+            status = None,
+            from_date = None,
+            to_date = None,
+            subject = None,
+            label = None,
+            intro = None,
+            note = None,
             tags = None,
-            supplier_id = None,
+            article_id = None,
+
             order_by = None,
         )
 
 
     def search(
         self,
-        article_number = None,
-        title = None,
-        description = None,
-        currency_code = None,
-        unit_id = None,
+        client_id = None,
+        contact_id = None,
+        invoice_number = None,
+        status = None,
+        from_date = None,
+        to_date = None,
+        subject = None,
+        label = None,
+        intro = None,
+        note = None,
         tags = None,
-        supplier_id = None,
+        article_id = None,
         order_by = None
     ):
         """
@@ -486,13 +608,19 @@ class RemindersIterator(ItemsIterator):
         """
 
         # Params
-        self.search_params.article_number = article_number
-        self.search_params.title = title
-        self.search_params.description = description
-        self.search_params.currency_code = currency_code
-        self.search_params.unit_id = unit_id
+        self.search_params.client_id = client_id
+        self.search_params.contact_id = contact_id
+        self.search_params.invoice_number = invoice_number
+        self.search_params.status = status
+        self.search_params.from_date = from_date
+        self.search_params.to_date = to_date
+        self.search_params.subject = subject
+        self.search_params.label = label
+        self.search_params.intro = intro
+        self.search_params.note = note
         self.search_params.tags = tags
-        self.search_params.supplier_id = supplier_id
+        self.search_params.article_id = article_id
+
         self.search_params.order_by = order_by
 
         # Search and prepare first page as result
@@ -502,13 +630,18 @@ class RemindersIterator(ItemsIterator):
     def load_page(self, page):
 
         self.items.search(
-            article_number = self.search_params.article_number,
-            title = self.search_params.title,
-            description = self.search_params.description,
-            currency_code = self.search_params.currency_code,
-            unit_id = self.search_params.unit_id,
+            client_id = self.search_params.client_id,
+            contact_id = self.search_params.contact_id,
+            invoice_number = self.search_params.invoice_number,
+            status = self.search_params.status,
+            from_date = self.search_params.from_date,
+            to_date = self.search_params.to_date,
+            subject = self.search_params.subject,
+            label = self.search_params.label,
+            intro = self.search_params.intro,
+            note = self.search_params.note,
             tags = self.search_params.tags,
-            supplier_id = self.search_params.supplier_id,
+            article_id = self.search_params.article_id,
             order_by = self.search_params.order_by,
 
             fetch_all = False,
@@ -517,5 +650,3 @@ class RemindersIterator(ItemsIterator):
             page = page,
             per_page = self.per_page
         )
-
-